requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
bcrypt==4.1.1
flask-cors==4.0.0
psycopg2-binary==2.9.9

